+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project are documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.2.1] - 2025-09-16
+
+### Added
+
+- `--outputType` Flag to configure output type. Defaults to auto.
+
+### Changed
+
+- Moved `--annotate` flag to `--outputType` 
+
 ## [0.2.0] - 2025-09-16
 
 ### Added